Added Backspace click and various other things
Backspace click is a little scuffed but it's as good as it's probably going to get. I tried adding enter but it's not a KeyboardEvent. Feel free to fork and use the audio I included if you figure out a way to do it.
Also added ZeresPluginLibrary and an auto-updater(that may or may not work), so when I change the sound files it doesn't break for so long.